+++ src/Request/RequestException.php
@@ -10,7 +10,7 @@
  * @author DrDelay <info@vi0lation.de>
  */
 
-namespace DrDelay\PokemonGo\Http;
+namespace DrDelay\PokemonGo\Request;
 
 use DrDelay\PokemonGo\PokemonGoException;